S7 ven 08 09 2021v1

+++ Realestate_portal/Controllers/CustomersController.cs
@@ -19,6 +19,8 @@ namespace Realestate_portal.Controllers
         private Realstate_agentsEntities db = new Realstate_agentsEntities();
         private clsGeneral generalClass = new clsGeneral();
         private Cls_GoogleCalendar cls_GoogleCalendar = new Cls_GoogleCalendar();
+      
+
         // GET: Tb_Customers/Details/5
         public ActionResult Details(int? id)
         {
@@ -134,7 +136,7 @@ namespace Realestate_portal.Controllers
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Create([Bind(Include = "ID_Customer,Name,LastName,Gender,Birthday,Marital_status,Type,Email,Phone,Mobile,State,Address,Zipcode,Lead,ID_User,Active,ID_Company,Creation_date,Source")] Tb_Customers tb_Customers)
+        public ActionResult Create([Bind(Include = "ID_Customer,Name,LastName,Gender,Birthday,Marital_status,Type,Email,Phone,Mobile,State,Address,Zipcode,Lead,Active,ID_Company,Creation_date,Source")] Tb_Customers tb_Customers)
         {
             Sys_Users activeuser = Session["activeUser"] as Sys_Users;
             if (tb_Customers.Birthday == null) { tb_Customers.Birthday = DateTime.UtcNow; }
@@ -143,9 +145,9 @@ namespace Realestate_portal.Controllers
             if (birthdemo < theDate) {
                 tb_Customers.Birthday = DateTime.UtcNow;
             }
-            var user_assigned = (from u in db.Sys_Users where (u.ID_User == tb_Customers.ID_User && u.Active == true) orderby u.LastName ascending select u).FirstOrDefault();
+            
         
-            tb_Customers.User_assigned = user_assigned.Name + " " + user_assigned.LastName;
+           
             tb_Customers.ID_Company = activeuser.ID_Company;
             tb_Customers.Lead = false;
             if (tb_Customers.Zipcode == null) { tb_Customers.Zipcode = ""; }
@@ -160,17 +162,10 @@ namespace Realestate_portal.Controllers
                 db.Tb_Customers.Add(tb_Customers);
                 db.SaveChanges();
 
-            Sys_Notifications newnotification = new Sys_Notifications();
-            newnotification.Active = true;
-            newnotification.Date = DateTime.UtcNow;
-            newnotification.Title = "New Customer assigned.";
-            newnotification.Description = "Customer: " + tb_Customers.Name + " " + tb_Customers.LastName + ".";
-            newnotification.ID_user = user_assigned.ID_User;
-            db.Sys_Notifications.Add(newnotification);
-            db.SaveChanges();
+           
 
 
-            return RedirectToAction("CustomerDashboard", "CRM", new { id = tb_Customers.ID_Customer, broker = 0});
+            return RedirectToAction("AssignList", "Tb_Customers_Users", new { id = tb_Customers.ID_Customer, broker = 0});
             
                 
             
@@ -193,8 +188,6 @@ namespace Realestate_portal.Controllers
                 newlead.Birthday = DateTime.UtcNow;
                 newlead.Marital_status = "";
                 newlead.Type = "";
-                newlead.ID_User = 4;
-                newlead.User_assigned = "Not Assigned";
                 newlead.State = "";
                 newlead.Lead = true;
                 newlead.ID_Company = 1;
@@ -441,7 +434,7 @@ namespace Realestate_portal.Controllers
         {
             var customer = (from a in db.Tb_Customers where (a.ID_Customer == id) select a).Include(f=> f.Sys_Company).FirstOrDefault();
 
-            var result = new { id = customer.ID_Customer, Name = customer.Name,Phone=customer.Phone, LastName = customer.LastName, Email = customer.Email, User = customer.User_assigned, DateCreated = customer.Creation_date };
+            var result = new { id = customer.ID_Customer, Name = customer.Name,Phone=customer.Phone, LastName = customer.LastName, Email = customer.Email, DateCreated = customer.Creation_date };
             
             return Json( result , JsonRequestBehavior.AllowGet);
         }
@@ -509,7 +502,7 @@ namespace Realestate_portal.Controllers
                                                       {
                                                           ID = t.ID_User,
                                                           FullName = t.Name + " " + t.LastName
-                                                      }), "ID", "FullName", tb_Customers.ID_User);
+                                                      }), "ID", "FullName");
                 }
                 else
                 {
@@ -525,7 +518,7 @@ namespace Realestate_portal.Controllers
                                                           {
                                                               ID = t.ID_User,
                                                               FullName = t.Name + " " + t.LastName
-                                                          }), "ID", "FullName", tb_Customers.ID_User);
+                                                          }), "ID", "FullName");
                     }
                     else
                     {
@@ -538,7 +531,7 @@ namespace Realestate_portal.Controllers
                                                               {
                                                                   ID = t.ID_User,
                                                                   FullName = t.Name + " " + t.LastName
-                                                              }), "ID", "FullName", tb_Customers.ID_User);
+                                                              }), "ID", "FullName");
                         }
                         else
                         {
@@ -587,7 +580,7 @@ namespace Realestate_portal.Controllers
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit([Bind(Include = "ID_Customer,Name,LastName,Gender,Birthday,Marital_status,Type,Email,Phone,Mobile,State,Address,Zipcode,Lead,ID_User,User_assigned,Active,ID_Company,Creation_date,Source")] Tb_Customers tb_Customers)
+        public ActionResult Edit([Bind(Include = "ID_Customer,Name,LastName,Gender,Birthday,Marital_status,Type,Email,Phone,Mobile,State,Address,Zipcode,Lead,Active,ID_Company,Creation_date,Source")] Tb_Customers tb_Customers)
         {
             try
             {               
@@ -600,40 +593,24 @@ namespace Realestate_portal.Controllers
                 Tb_Customers customer = (from a in db.Tb_Customers.Where(a=> a.ID_Customer==tb_Customers.ID_Customer) select a ).AsNoTracking().FirstOrDefault();
                 tb_Customers.Sys_Company = db.Sys_Company.Find(tb_Customers.ID_Company);
                 tb_Customers.Tb_Process = (from a in db.Tb_Process.Where(a => a.ID_Customer == tb_Customers.ID_Customer) select a).ToList();
-                var usuarioantes = db.Tb_Customers.Where(c => c.ID_Customer == tb_Customers.ID_Customer).Select(c => c.ID_User).FirstOrDefault();
+               
 
                 db.Entry(tb_Customers).State=EntityState.Modified;
                 db.SaveChanges();
                 TempData["exito"] = "Customer info updated successfully.";
 
-                if (usuarioantes != tb_Customers.ID_User)
-                {
-                    var customerislead = (from h in db.Tb_Customers where (h.ID_Customer == tb_Customers.ID_Customer) select h).FirstOrDefault();
-
-                    if (tb_Customers.ID_User == 4)
-                    {
-                        customerislead.Lead = true;
-                        db.Entry(customerislead).State = EntityState.Modified;
-                        db.SaveChanges();
-                    }
-                    else {
-                        if (customerislead.Lead == true)
-                        {
-                            customerislead.Lead = false;
-                            db.Entry(customerislead).State = EntityState.Modified;
-                            db.SaveChanges();
-                        }
-                    }
+              
+                       
            
-                    Sys_Notifications newnotification = new Sys_Notifications();
-                    newnotification.Active = true;
-                    newnotification.Date = DateTime.UtcNow;
-                    newnotification.Title = "New Customer assigned.";
-                    newnotification.Description = "Customer: " + tb_Customers.Name + " " + tb_Customers.LastName + ".";
-                    newnotification.ID_user = tb_Customers.ID_User;
-                    db.Sys_Notifications.Add(newnotification);
-                    db.SaveChanges();                                       
-                }
+                    //Sys_Notifications newnotification = new Sys_Notifications();
+                    //newnotification.Active = true;
+                    //newnotification.Date = DateTime.UtcNow;
+                    //newnotification.Title = "New Customer assigned.";
+                    //newnotification.Description = "Customer: " + tb_Customers.Name + " " + tb_Customers.LastName + ".";
+                    //newnotification.ID_user = tb_Customers.ID_User;
+                    //db.Sys_Notifications.Add(newnotification);
+                    //db.SaveChanges();                                       
+               
 
                 return RedirectToAction("CustomerDashboard", "CRM", new {id=tb_Customers.ID_Customer, broker= 0 });
             }
@@ -746,11 +723,24 @@ namespace Realestate_portal.Controllers
             var result="";
             try
             {
-                Tb_Customers tb_Customers = db.Tb_Customers.Find(id);
-                db.Tb_Customers.Remove(tb_Customers);
-                db.SaveChanges();
+            
+               
+                Tb_Customers_UsersController tb_Customers_Users = new Tb_Customers_UsersController();
+                var delete_Custo = tb_Customers_Users.Delete(id);
+                if (delete_Custo)
+                {
+                    Tb_Customers tb_Customers = db.Tb_Customers.Find(id);
+                    db.Tb_Customers.Remove(tb_Customers);
+                    db.SaveChanges();
 
-                 result = "SUCCESS";
+                    result = "SUCCESS";
+                }
+                else {
+
+                    result = "ERROR";
+                }
+                
+                
             }
             catch(Exception EX)
             {
